Merge revisions in original function
Merge all the edits on the one-off scoring method to the original function score_drugs_vs_control instead of writing an extra function. Goal is to provide smooth switch between versions without changing the code in programs that calls Orobas. This branch should serve as an alternative in the short term.